+++ src/Acl/ResourcePermissions.php
@@ -17,6 +17,7 @@ use Konekt\Acl\PermissionRegistrar;
 
 /**
  * Utility class for handling standard resource permissions
+ * @deprecated
  */
 class ResourcePermissions
 {