+++ tests/Dummies/IssueTypeController.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the IssueTypeController class.
+ *
+ * @copyright   Copyright (c) 2020 Attila Fulop
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2020-09-27
+ *
+ */
+
+namespace Konekt\AppShell\Tests\Dummies;
+
+class IssueTypeController
+{
+    public function create()
+    {
+        return "Hello";
+    }
+}